refactor: use txWrappers factory
